Refactor different frameworks into their own frameworks folder
