+++ blankly/deployment/reporter_headers.py
@@ -22,8 +22,8 @@ import ssl
 from typing import Any
 
 from blankly.utils.utils import load_notify_preferences
-from blankly.strategy.strategy_base import Strategy
-from blankly.strategy.signal import Signal
+from blankly.framework.strategy import Strategy
+from blankly.framework.signal.signal import Signal
 
 
 class Reporter:
